Move persistent lifecycle
