+++ packages/common/async/src/index.ts
@@ -12,6 +12,7 @@ export * from './latch';
 export * from './observable';
 export * from './observable-value';
 export * from './mutex';
+export * from './persistent-lifecycle';
 export * from './sink';
 export * from './stream-to-array';
 export * from './timeout';
@@ -19,8 +20,8 @@ export * from './timer';
 export * from './testing';
 export * from './trigger';
 export * from './types';
-export * from './until';
 export * from './task-scheduling';
 export * from './test-stream';
 export * from './track-leaks';
+export * from './until';
 export * from './update-scheduler';
