+++ packages/common/context/src/index.ts
@@ -4,6 +4,6 @@
 
 export * from './context';
 export * from './context-disposed-error';
-export * from './persistent-lifecycle';
+export * from '../../async/src/persistent-lifecycle';
 export * from './promise-utils';
 export * from './resource';
